requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
aioredis==2.0.1
aiofiles===23.2.1
anthropic==0.39.0
async-timeout==4.0.3
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
chatgpt-md-converter==0.1.2
firebase-admin==6.6.0
google-auth==2.36.0
google-api-core==2.23.0
google-cloud-bigquery==3.27.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
googleapis-common-protos==1.66.0
google-generativeai==0.8.3
gcloud-aio-auth==5.3.2
gcloud-aio-storage==9.3.0
grpcio==1.68.0
idna==3.10
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycryptodome==3.21.0
PyJWT==2.10.0
pillow==10.4.0
redis==5.2.0
requests==2.32.3
replicate==1.0.3
runwayml==2.1.0
six==1.16.0
stripe==11.3.0
openai==1.55.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydub==0.25.1
pytz==2024.2
uvicorn==0.32.1
fastapi==0.115.5
ffmpeg==1.4
ffprobe-python==1.0.3
yookassa==3.4.1
